Changed: bump tree-sitter to 0.27
- Update tree-sitter from 0.26 to 0.27; tree-sitter-language 0.1.7 to 0.1.8.
- Adapt parent_field_name helpers to the new borrowed-lifetime
  field_name_for_child return type.
- Full workspace tests pass on the new version: 641 tests, 0 failures.

+++ src/rust-llm-tidy/src/languages/csharp/parse.rs
@@ -350,7 +350,7 @@ fn build_members(
         .windows(2)
         .any(|pair| pair[0].end_position().row >= pair[1].start_position().row)
         || decls.last().is_some_and(|last| {
-            list.child(list.child_count().saturating_sub(1) as u32)
+            list.child(list.child_count().saturating_sub(1))
                 .is_some_and(|brace| brace.start_position().row <= last.end_position().row)
         })
         || decls.first().is_some_and(|first| {
